+++ manager-api/src/api/notifications/notifications.module.ts
@@ -1,9 +1,10 @@
 import { Module } from "@nestjs/common";
+import { CustomPermissionGuardModule } from "../../core/custom-permission-guard/custom-permission-guard.module";
 import { NotificationsModule as CoreNotificationsModule } from "../../core/notifications/notifications.module";
 import { NotificationsController } from "./notifications.controller";
 
 @Module({
-  imports: [CoreNotificationsModule],
+  imports: [CoreNotificationsModule, CustomPermissionGuardModule],
   controllers: [NotificationsController],
 })
 export class NotificationsApiModule {}